Revert "minor #32054 Prepare for PHP 7.4 preload (nicolas-grekas)"
This reverts commit a0aa94114a7b0334f94eb53c7536c729c1be2fef, reversing
changes made to 8496003634906d220ee1bc95f1547ced64545ffa.
