+++ src/Symfony/Component/Cache/Exception/CacheException.php
@@ -11,10 +11,15 @@
 
 namespace Symfony\Component\Cache\Exception;
 
+use Psr\Cache\CacheException as Psr6CacheInterface;
 use Psr\SimpleCache\CacheException as SimpleCacheInterface;
 
 if (interface_exists(SimpleCacheInterface::class)) {
-    require __DIR__.\DIRECTORY_SEPARATOR.'CacheException+psr16.php';
+    class CacheException extends \Exception implements Psr6CacheInterface, SimpleCacheInterface
+    {
+    }
 } else {
-    require __DIR__.\DIRECTORY_SEPARATOR.'CacheException-psr16.php';
+    class CacheException extends \Exception implements Psr6CacheInterface
+    {
+    }
 }
