--- src/Symfony/Component/Cache/Exception/InvalidArgumentException+psr16.php
@@ -1,19 +1,0 @@
-<?php
-
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Symfony\Component\Cache\Exception;
-
-use Psr\Cache\InvalidArgumentException as Psr6CacheInterface;
-use Psr\SimpleCache\InvalidArgumentException as SimpleCacheInterface;
-
-class InvalidArgumentException extends \InvalidArgumentException implements Psr6CacheInterface, SimpleCacheInterface
-{
-}
