+++ src/Symfony/Component/Cache/Exception/InvalidArgumentException.php
@@ -11,10 +11,15 @@
 
 namespace Symfony\Component\Cache\Exception;
 
+use Psr\Cache\InvalidArgumentException as Psr6CacheInterface;
 use Psr\SimpleCache\InvalidArgumentException as SimpleCacheInterface;
 
 if (interface_exists(SimpleCacheInterface::class)) {
-    require __DIR__.\DIRECTORY_SEPARATOR.'InvalidArgumentException+psr16.php';
+    class InvalidArgumentException extends \InvalidArgumentException implements Psr6CacheInterface, SimpleCacheInterface
+    {
+    }
 } else {
-    require __DIR__.\DIRECTORY_SEPARATOR.'InvalidArgumentException-psr16.php';
+    class InvalidArgumentException extends \InvalidArgumentException implements Psr6CacheInterface
+    {
+    }
 }
