+++ src/Symfony/Component/Cache/Exception/LogicException.php
@@ -11,10 +11,15 @@
 
 namespace Symfony\Component\Cache\Exception;
 
-use Psr\SimpleCache\LogicException as SimpleCacheInterface;
+use Psr\Cache\CacheException as Psr6CacheInterface;
+use Psr\SimpleCache\CacheException as SimpleCacheInterface;
 
 if (interface_exists(SimpleCacheInterface::class)) {
-    require __DIR__.\DIRECTORY_SEPARATOR.'LogicException+psr16.php';
+    class LogicException extends \LogicException implements Psr6CacheInterface, SimpleCacheInterface
+    {
+    }
 } else {
-    require __DIR__.\DIRECTORY_SEPARATOR.'LogicException-psr16.php';
+    class LogicException extends \LogicException implements Psr6CacheInterface
+    {
+    }
 }
